+++ larcoh/configs/prophecy-test.toml
@@ -1,0 +1,22 @@
+[workflow]
+access_level = 'test'
+dataset = 'prophecy'
+datasets = ['prophecy']
+dataset_gcp_project = 'prophecy-339301'
+check_expected_outputs = true
+skip_samples_with_missing_input = true
+sequencing_type = 'genome'
+skip_samples = [
+  'CPG216275',  # https://batch.hail.populationgenomics.org.au/batches/74382
+  'CPG216010',  # https://batch.hail.populationgenomics.org.au/batches/74382
+]
+scatter_count = 20
+output_version = '0.1'
+vds_version = '0.1'
+
+[dataproc]
+highmem_workers = false
+
+[hail]
+billing_project = 'prophecy'
+dataproc_cluster_id = 'dp-82ffa0cea12545be87f2'
