--- large_cohort/configs/prophecy-test.toml
@@ -1,50 +1,0 @@
-[workflow]
-access_level = 'test'
-dataset = 'prophecy'
-datasets = ['prophecy']
-dataset_gcp_project = 'prophecy-339301'
-check_expected_outputs = true
-skip_samples_with_missing_input = true
-sequencing_type = 'genome'
-skip_samples = [
-  'CPG216275',  # https://batch.hail.populationgenomics.org.au/batches/74382
-  'CPG216010',  # https://batch.hail.populationgenomics.org.au/batches/74382
-]
-scatter_count = 20
-output_version = '0.1'
-vds_version = '0.1'
-
-[dataproc]
-highmem_workers = false
-
-[hail]
-pool_label = 'large-cohort'
-delete_scratch_on_exit = false
-billing_project = 'prophecy'
-dataproc_cluster_id = 'dp-82ffa0cea12545be87f2'
-
-[larhoc.hardfiltering]
-min_coverage = 18
-max_n_snps = 8000000
-min_n_snps = 2400000
-max_n_singletons = 800000
-max_r_duplication = 0.3
-max_r_het_hom = 3.3
-max_r_contamination = 0.04
-min_insert_size = 250
-
-[larhoc.pca]
-# maximum kin threshold to be considered unrelated
-max_kin = 0.2
-# minimum Random Forest probability for population assignment
-min_pop_prob = 0.5
-
-[larhoc.vqsr]
-# VQSR, when applying model, targets indel_filter_level and snp_filter_level
-# sensitivities. The tool matches them internally to a VQSLOD score cutoff 
-# based on the model's estimated sensitivity to a set of true variants.
-snp_filter_level = 99.7
-indel_filter_level = 99.0
-
-[references]
-ancestry_ht = 'ancestry/v3-90k/pca_sites.ht'
